+++ src/utils/GatsbyanUtils.tsx
@@ -12,13 +12,13 @@ export const kebabCase = (str: string) => {
   return slugify(str);
 };
 
-export const getPublishDate = (date: string | Date) => formatToPattern(date, "MMMM do, yyyy");
+export const getPublishDate = (date: string) => formatToPattern(date, "MMMM do, yyyy");
 
-export const getPublishDateTime = (date: string | Date) => formatToPattern(date, "eee. MMM do, yyyy hh:mm a");
+export const getPublishDateTime = (date: string) => formatToPattern(date, "eee. MMM do, yyyy hh:mm a");
 
-export const getDateYear = (date: string | Date) => formatToPattern(date, "yyyy-MM-dd");
+export const getDateYear = (date: string) => formatToPattern(date, "yyyy-MM-dd");
 
-export const getMonthYearDate = (date: string | Date) => formatToPattern(date, "yyyy-MMMM");
+export const getMonthYearDate = (date: string) => formatToPattern(date, "yyyy-MMMM");
 
 export const toNow = (date: string | Date) => formatDistanceToNow(date);
 
@@ -26,8 +26,8 @@ export const isAfterDate = (date1: string | Date, date2: string | Date) => isAft
 
 export const plusDays = (date: string | Date, day: number) => add(date, { days: day });
 
-function formatToPattern(dateArg: string | Date, formatString: string, timeZone: string = "Asia/Jakarta") {
-  const date = typeof dateArg === "string" ? new Date(dateArg) : dateArg;
+function formatToPattern(dateArg: string, formatString: string, timeZone: string = "Asia/Jakarta") {
+  const date = new Date(dateArg);
   const zonedDate = date.toLocaleString("en-US", { timeZone: timeZone });
   return format(zonedDate, formatString);
 }
@@ -91,28 +91,6 @@ export const PAGE_COUNT = 5;
 export const SEARCH_COUNT = 7;
 
 export const DEFAULT_ICON_SIZE = 20;
-
-export function isCommentShown() {
-  const [commentShown, setCommentShown] = React.useState(false);
-
-  React.useEffect(() => {
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
-  const handleScroll = () => {
-    const innerHeight = window.innerHeight + document.documentElement.scrollTop;
-    const clientHeight = document.body.clientHeight;
-    const percentage = (innerHeight / clientHeight) * 100;
-    if (!commentShown && percentage > 50) {
-      setCommentShown(true);
-    }
-  };
-  return commentShown;
-}
 
 export const getPagingCalculator = (currentPage: number, totalPage: number): PaginationProp => {
   const range = Math.floor(PAGE_COUNT / 2);
